requirements.txt
tweepy
textblob
python-dotenv